refactor: move LabelledCheckbox to it's own file
